Minor refactor and some fixes
